// src/app/components/votes-admin/models/models.ts
export interface Models {
}

export interface Enterprise {
    _id: string;
    address: string;
    city: string;
    code: string;
    contry: string;
    cuota: number;
    email: string;
    name: string;
    nif: string;
    section: string;
    status: boolean;
    total_votes: number;
    type_vote: string;
    vote_delegate: number;
}